Use 64-bit unsigned integer for bounding box.

+++ branches/dataspaces-epsi/include/sfc.h
@@ -35,9 +35,9 @@
 #define _hilbert_h_
 
 /* define the bitmask_t type as an integer of sufficient size */
-typedef unsigned long long bitmask_t;
+typedef uint64_t bitmask_t;
 /* define the halfmask_t type as an integer of 1/2 the size of bitmask_t */
-typedef unsigned long halfmask_t;
+typedef uint32_t halfmask_t;
 
 #define adjust_rotation(rotation,nDims,bits)                            \
 do {                                                                    \
